Azure Function Code/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy==1.24.3
scikit-learn==1.2.2
pandas==2.0.1